Nicolas Regnault: - add 3 band Hilbert space - add code to apply inversio symmetry (not yet working) - still debugging FQHEQuantumSpinHall2DSimpleTI
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1955 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/FermionOnSphereWithSU3Spin.cc
@@ -642,7 +642,7 @@ long FermionOnSphereWithSU3Spin::ShiftedEvaluateHilbertSpaceDimension(int nbrFer
 // totalY = three time the total Y value
 // return value = Hilbert space dimension
 
-int FermionOnSphereWithSU3Spin::EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalTz, int totalY)
+long FermionOnSphereWithSU3Spin::EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalTz, int totalY)
 {
   int N1 = (2 * nbrFermions) + totalY + (3 * totalTz);
   int N2 = (2 * nbrFermions) + totalY - (3 * totalTz);
